Different colors for different ball types
This is something I worked out a couple of years ago. Considering the nature and scope of the project, I think it fits.

The code changes the color of Poke, Great, Ultra, Master, and Safari balls during their throw animation. Pokeballs are the default red, Greats are blue, Ultras are grey, Masters are purple and Safaris are green. 

Just adds a bit more color to the game.